backend/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contextlib2==21.6.0
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.61.1
django-cors-headers==4.3.1
django-ninja==1.1.0
django-ninja-crud==0.4.1
django-ninja-extra==0.20.5
django-ninja-jwt==5.3.0
idna==3.6
injector==0.21.0
oauthlib==3.2.2
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
cachetools==5.3.3 
google-auth==2.29.0 
pyasn1==0.6.0 
pyasn1-modules==0.4.0 
rsa==4.9
uvicorn==0.29.0
